{"schema":"decodex/commit/1","summary":"Move downward candidate selection","authority":"manual"}
