+++ packages/rsnap-overlay/src/scroll_capture/support.rs
@@ -10,11 +10,9 @@ use image::{
 use crate::scroll_capture::OverlapMatch;
 use crate::scroll_capture::{
 	DIRECTION_WARNING_MARGIN_X100, DOWNWARD_REGISTRATION_AMBIGUOUS_GAP_ROWS,
-	DOWNWARD_REGISTRATION_MIN_OVERLAP_DIVISOR, DOWNWARD_VIEWPORT_AUTHORITY_GAP_ROWS,
-	DirectionMatch, DownwardRegistration, DownwardViewportCandidate,
-	DownwardViewportCandidateSource, DownwardViewportResolution, InformativeSpan,
-	OverlapSearchConfig, RESUME_DIRECT_PROOF_MAX_MEAN_ABS_DIFF_X100, ScrollDirection,
-	ScrollFrameFingerprint, ScrollObserveOutcome,
+	DOWNWARD_REGISTRATION_MIN_OVERLAP_DIVISOR, DirectionMatch, DownwardRegistration,
+	InformativeSpan, OverlapSearchConfig, RESUME_DIRECT_PROOF_MAX_MEAN_ABS_DIFF_X100,
+	ScrollDirection, ScrollFrameFingerprint, ScrollObserveOutcome,
 };
 
 const INFORMATIVE_SPAN_ROW_SAMPLES: u32 = 24;
@@ -124,86 +122,6 @@ pub(crate) fn compose_provisional_preview_image(
 	let preview_strip = resize_strip_to_preview_width(&strip, preview_width_px);
 
 	append_vertical_image(base_preview, &preview_strip).unwrap_or_else(|_| base_preview.clone())
-}
-
-pub(super) fn select_downward_viewport_candidate(
-	candidates: &mut [DownwardViewportCandidate],
-) -> DownwardViewportResolution {
-	if candidates.is_empty() {
-		return DownwardViewportResolution::NoMatch;
-	}
-
-	if let Some(preferred_local) = prefer_local_downward_viewport_candidate(candidates) {
-		let ambiguity_margin = downward_viewport_competing_margin(candidates, preferred_local);
-		let competing = candidates.iter().copied().find(|candidate| {
-			candidate != &preferred_local
-				&& candidate.viewport_top_y.abs_diff(preferred_local.viewport_top_y)
-					>= DOWNWARD_VIEWPORT_AUTHORITY_GAP_ROWS
-				&& candidate.mean_abs_diff_x100
-					<= preferred_local.mean_abs_diff_x100.saturating_add(ambiguity_margin)
-		});
-
-		return match competing {
-			Some(competing) => {
-				DownwardViewportResolution::Ambiguous { preferred: preferred_local, competing }
-			},
-			None => DownwardViewportResolution::Selected(preferred_local),
-		};
-	}
-
-	candidates.sort_by(|left, right| {
-		left.mean_abs_diff_x100
-			.cmp(&right.mean_abs_diff_x100)
-			.then(left.source.priority().cmp(&right.source.priority()))
-			.then(left.motion_rows.cmp(&right.motion_rows))
-	});
-
-	let preferred = candidates[0];
-	let ambiguity_margin = downward_viewport_competing_margin(candidates, preferred);
-	let competing = candidates.iter().copied().skip(1).find(|candidate| {
-		candidate.viewport_top_y.abs_diff(preferred.viewport_top_y)
-			>= DOWNWARD_VIEWPORT_AUTHORITY_GAP_ROWS
-			&& candidate.mean_abs_diff_x100
-				<= preferred.mean_abs_diff_x100.saturating_add(ambiguity_margin)
-	});
-
-	match competing {
-		Some(competing) => DownwardViewportResolution::Ambiguous { preferred, competing },
-		None => DownwardViewportResolution::Selected(preferred),
-	}
-}
-
-pub(super) fn format_downward_viewport_candidates(
-	candidates: &[DownwardViewportCandidate],
-) -> String {
-	candidates
-		.iter()
-		.map(|candidate| {
-			format!(
-				"{:?}@{}/{}:{}",
-				candidate.source,
-				candidate.viewport_top_y,
-				candidate.motion_rows,
-				candidate.mean_abs_diff_x100
-			)
-		})
-		.collect::<Vec<_>>()
-		.join(",")
-}
-
-pub(super) fn best_local_downward_viewport_candidate(
-	candidates: &[DownwardViewportCandidate],
-) -> Option<DownwardViewportCandidate> {
-	candidates
-		.iter()
-		.copied()
-		.filter(|candidate| candidate.source != DownwardViewportCandidateSource::CommittedKeyframe)
-		.min_by(|left, right| {
-			left.mean_abs_diff_x100
-				.cmp(&right.mean_abs_diff_x100)
-				.then(left.source.priority().cmp(&right.source.priority()))
-				.then(left.motion_rows.cmp(&right.motion_rows))
-		})
 }
 
 pub(super) fn evaluate_overlap_direction(
@@ -622,46 +540,6 @@ pub(super) fn evenly_spaced_sample(start: u32, end_exclusive: u32, index: u32, c
 		(u64::from(index) * u64::from(span.saturating_sub(1))) / u64::from(count.saturating_sub(1));
 
 	start.saturating_add(numerator as u32).min(end_exclusive.saturating_sub(1))
-}
-
-fn downward_viewport_competing_margin(
-	candidates: &[DownwardViewportCandidate],
-	preferred: DownwardViewportCandidate,
-) -> u32 {
-	let exact_corroborated = candidates.iter().any(|candidate| {
-		candidate != &preferred
-			&& candidate.viewport_top_y == preferred.viewport_top_y
-			&& candidate.motion_rows == preferred.motion_rows
-			&& candidate.mean_abs_diff_x100
-				<= preferred.mean_abs_diff_x100.saturating_add(DIRECTION_WARNING_MARGIN_X100)
-	});
-
-	if exact_corroborated { 0 } else { DIRECTION_WARNING_MARGIN_X100 }
-}
-
-fn prefer_local_downward_viewport_candidate(
-	candidates: &[DownwardViewportCandidate],
-) -> Option<DownwardViewportCandidate> {
-	let local = best_local_downward_viewport_candidate(candidates)?;
-	let committed = candidates
-		.iter()
-		.copied()
-		.filter(|candidate| candidate.source == DownwardViewportCandidateSource::CommittedKeyframe)
-		.min_by(|left, right| {
-			left.mean_abs_diff_x100
-				.cmp(&right.mean_abs_diff_x100)
-				.then(left.motion_rows.cmp(&right.motion_rows))
-		});
-	let Some(committed) = committed else {
-		return Some(local);
-	};
-	let committed_is_nearby = committed.viewport_top_y.abs_diff(local.viewport_top_y)
-		< DOWNWARD_VIEWPORT_AUTHORITY_GAP_ROWS;
-	let committed_is_only_modestly_better =
-		committed.mean_abs_diff_x100.saturating_add(DIRECTION_WARNING_MARGIN_X100)
-			>= local.mean_abs_diff_x100;
-
-	if committed_is_nearby && committed_is_only_modestly_better { Some(local) } else { None }
 }
 
 #[cfg(test)]
